sdk: split query_verbs into module directory (4/9: extract compare)
Move compare, compute_compare_excluded, and helpers to query_verbs/compare.rs.
Promote normalize_provider_filter to shared plumbing in mod.rs (used by
both compare and hotspots).
